% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesij_lib.R
\name{ComputeIJCovariance}
\alias{ComputeIJCovariance}
\title{Return an estimate of the infinitesimal jackknife covariance estimate
of the frequentist variance of the posterior expectations of the parameters
in param_draws_mat.}
\usage{
ComputeIJCovariance(loglik_draws_mat, param_draws_mat, exchangeable_col = NULL)
}
\arguments{
\item{loglik_draws_mat}{Draws of the log likelihood.}

\item{param_draws_mat}{Draws of the parameters of interest..}
}
\value{
The IJ covariance matrix, which is an estimate of
N * Cov(E[params | x]), where N is the number of distinct exchangeable
observations.
}
\description{
Return an estimate of the infinitesimal jackknife covariance estimate
of the frequentist variance of the posterior expectations of the parameters
in param_draws_mat.
}
