% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers_lib.R
\name{GatherRowNamedMatrix}
\alias{GatherRowNamedMatrix}
\title{Gather a matrix as with \code{gather_draws}, adding a column for the
rownames.}
\usage{
GatherRowNamedMatrix(mat, ...)
}
\description{
Gather a matrix as with \code{gather_draws}, adding a column for the
rownames.
}
