% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_standard_errors_lib.R
\name{GetBlockBootstrapStandardErrors}
\alias{GetBlockBootstrapStandardErrors}
\title{Estimate Monte Carlo standard errors of sample covariances or
covariance-like functions by block bootstrapping the MCMC chain.}
\usage{
GetBlockBootstrapStandardErrors(
  draws1_mat,
  draws2_mat,
  num_blocks,
  num_draws,
  cov_fun = cov,
  show_progress_bar = FALSE
)
}
\arguments{
\item{draws1_mat}{One set of parameter draws.}

\item{draws2_mat}{Another set of parameter draws.}

\item{num_blocks}{The number of blocks in the block bootstrap.}

\item{num_draws}{The number of bootstrap draws.}

\item{cov_fun}{Optional.  A function of draws1_mat and draws2_mat which
you want to bootstrap.  By default, the covariance is computed.}

\item{show_progress_par.}{Optional.  If TRUE, show a progress bar.
By default, FALSE.}
}
\value{
A list containing the draws of the output of cov_fun in cov_samples
and the estimated Monte Carlo sample errors in cov_se.
}
\description{
Estimate Monte Carlo standard errors of sample covariances or
covariance-like functions by block bootstrapping the MCMC chain.
}
