% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetHyperparameterDataFrame}
\alias{GetHyperparameterDataFrame}
\title{Get a dataframe with the hyperparameter values contained in a stan data list.}
\usage{
GetHyperparameterDataFrame(stan_sensitivity_list, stan_data)
}
\arguments{
\item{stan_sensitivity_list}{The output of \code{GetStanSensitivityModel}}

\item{stan_data}{The Stan data for the original model.}
}
\value{
A tidy dataframe with the hyperparameter names and values.
}
\description{
Get a dataframe with the hyperparameter values contained in a stan data list.
}
