% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetImportanceSamplingFromModelFit}
\alias{GetImportanceSamplingFromModelFit}
\title{Use importance sampling to approximate means at different hyperparameter
values.  Experimental use only.}
\usage{
GetImportanceSamplingFromModelFit(
  sampling_result,
  draws_mat,
  stan_sensitivity_list,
  imp_sens_par_list,
  lp_vec = NULL
)
}
\description{
Use importance sampling to approximate means at different hyperparameter
values.  Experimental use only.
}
