% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers_lib.R
\name{GetMCMCDataFrame}
\alias{GetMCMCDataFrame}
\title{Convert a Stan sampling result to a form that can be joined with tidy
sensitivity results.}
\usage{
GetMCMCDataFrame(sampling_result, cols = c("mean", "se_mean", "sd",
  "n_eff", "Rhat"))
}
\arguments{
\item{sampling_result}{The output of \code{stan::sampling}}

\item{cols}{The columns of the summary to keep.}
}
\value{
A data frame with the MCMC reuslts that can be joined with tidy
sensitivity results.
}
\description{
Convert a Stan sampling result to a form that can be joined with tidy
sensitivity results.
}
