% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetSamplingModelFilename}
\alias{GetSamplingModelFilename}
\title{Get the filename of the stan model to be used for sampling.}
\usage{
GetSamplingModelFilename(model_name)
}
\arguments{
\item{model_name}{A full path to the base model name.
For example, if the model with the \code{hyperparameter} block is named
\code{/home/user/my_model.stan}, set \code{model_name} should be equal
to \code{"/home/user/my_model"}}
}
\value{
The full path of the generated file to be used for sampling.
}
\description{
Get the filename of the stan model to be used for sampling.
}
\examples{
model_name <- GenerateSensitivityFromModel("models/my_model.stan")
model <- stan_model(GetSamplingModelFilename(model_name))
sampling_result <- sampling(model, data=stan_data)
}
