% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_processing_lib.R
\name{GetSensitivityFromGrads}
\alias{GetSensitivityFromGrads}
\title{Get a sensitivity matrix from a matrix of partial derivatives and
a matrix of draws.}
\usage{
GetSensitivityFromGrads(grad_mat, draws_mat)
}
\arguments{
\item{grad_mat}{A matrix of gradients of the log probability with respect
to hyperparameters.  The rows are hyperparameters, and the columns are
draws.}

\item{draws_mat}{A matrix of draws of parameters, the means of which
are the quantities of interest.  The rows are draws, and the columns
are parameters.}
}
\value{
A matrix of estimated derivatives, dE[parameter | hyper] / dhyper.
  The rows are hyperparameters and the columns are parameters.
}
\description{
Get a sensitivity matrix from a matrix of partial derivatives and
a matrix of draws.
}
