% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_standard_errors_lib.R
\name{GetSensitivityStandardErrors}
\alias{GetSensitivityStandardErrors}
\title{Estimate Monte Carlo standard errors for the sensitivity matrices using
a normal approximation and the delta method.  These standard errors are
based on what may be unrealistic assumptions, and should be interpreted with
caution.}
\usage{
GetSensitivityStandardErrors(draws_mat, grad_mat, fix_mean = FALSE,
  normalized = FALSE)
}
\arguments{
\item{draws_mat}{Parameter draws from \code{GetStanSensitivityFromModelFit}.}

\item{grad_mat}{Gradients from \code{GetStanSensitivityFromModelFit}.}

\item{fix_mean}{Experimental: estimate standard errors as if the means were
known exactly.  Not recommended.}

\item{normalized}{Compute standard errors for the normalized sensitivities.}
}
\value{
A matrix of the same dimensions as \code{sens_mat} containing
estimates of the Monte Carlo standard errors of the sensitivities.
}
\description{
Estimate Monte Carlo standard errors for the sensitivity matrices using
a normal approximation and the delta method.  These standard errors are
based on what may be unrealistic assumptions, and should be interpreted with
caution.
}
