% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetStanSensitivityFromModelFit}
\alias{GetStanSensitivityFromModelFit}
\title{Process the results of Stan samples and GetStanSensitivityModel into a
sensitivity matrix.}
\usage{
GetStanSensitivityFromModelFit(sampling_result, stan_sensitivity_list)
}
\arguments{
\item{sampling_result}{The output of \code{stan::sampling}}

\item{stan_sensitivity_list}{The output of \code{GetStanSensitivityModel}}
}
\value{
A list of matrices.  The elements of the list are
\itemize{
    \item{sens_mat: }{The local sensitivity of each posterior parameter to each hyperparameter.}
    \item{sens_mat_normalized: }{The same quantities as \code{sens_mat}, but normalized by the posterior standard deviation.}
    \item{grad_mat: }{The gradients of the log posterior evaluatated at the draws.}
    \item{lp_vec: }{The log probability of the model at each draw.}
    \item{draws_mat: }{The parameter draws in the same order as that of \code{grad_mat}}.
}
The result can be used directly, or passed to \code{GetTidyResult}.
}
\description{
Process the results of Stan samples and GetStanSensitivityModel into a
sensitivity matrix.
}
