% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetStanSensitivityModel}
\alias{GetStanSensitivityModel}
\title{Compile the sensitivity model and get a stanfit object and related
information.}
\usage{
GetStanSensitivityModel(model_name, stan_data)
}
\arguments{
\item{model_name}{A full path to the base model name.
For example, if the model with the \code{hyperparameter} block is named
\code{/home/user/my_model.stan}, set \code{model_name} should be equal
to \code{"/home/user/my_model"}}

\item{stan_data}{The Stan data list (e.g., for the \code{data} argument of
\code{stan::sampling})}
}
\value{
A list of Stan models and parameters names which can be passed to
the \code{stan_sensitivity_list} argument of
functions like code{GetStanSensitivityFromModelFit}.
}
\description{
Compile the sensitivity model and get a stanfit object and related
information.
}
