% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesij_lib.R
\name{GroupLogLikelihoodDraws}
\alias{GroupLogLikelihoodDraws}
\title{Aggregate the draws of the log likelihood according to a grouping variable
where the observations are exchangeable within a group.}
\usage{
GroupLogLikelihoodDraws(loglik_draws_mat, exchangeable_col)
}
\arguments{
\item{loglik_draws_mat}{Draws of the log likelihood.}

\item{exchangeable_col}{Optional, a vector of indices indicating the grouping
of exchangeable observations.}
}
\value{
The log likelihood draws of aggregated by the given grouping.
}
\description{
Aggregate the draws of the log likelihood according to a grouping variable
where the observations are exchangeable within a group.
}
