% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers_lib.R
\name{LoadStanData}
\alias{LoadStanData}
\title{Execute an R file and store its environment variables in a list.}
\usage{
LoadStanData(data_file)
}
\arguments{
\item{data_file}{The filename of a code containing R code..}
}
\value{
A list with the variables defined when running `data_file`.
}
\description{
Execute an R file and store its environment variables in a list.
}
