% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_processing_lib.R
\name{PlotSensitivities}
\alias{PlotSensitivities}
\title{Make a barchart from the output of \code{GetTidyResult}.}
\usage{
PlotSensitivities(sens_df, normalized = TRUE, se_num = 2)
}
\arguments{
\item{sens_df}{The output of \code{GetTidyResult}.}

\item{normalized}{Whether to display the normalized sensitivities.}

\item{se_num}{The number of standard errors to plot with the errorbars.}
}
\value{
A \code{ggplot} object showing the sensitivites of each parameter
to each hyperparameter.
}
\description{
Make a barchart from the output of \code{GetTidyResult}.
}
