% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers_lib.R
\name{PredictSensitivityFromStanData}
\alias{PredictSensitivityFromStanData}
\title{Use the linear approximation to predict the sensitivity to a new
stan data list.}
\usage{
PredictSensitivityFromStanData(stan_sensitivity_list, sens_result,
  stan_data, stan_data_perturb, description = "perturbation")
}
\arguments{
\item{stan_sensitivity_list}{The output of \code{GetStanSensitivityModel}}

\item{stan_data}{The original stan data at which
\code{stan_sensitivity_list} was calculated.}

\item{stan_data_perturb}{A new stan data file with different hyperparameters.}

\item{description}{A hyperparameter name to describe this perturbation.}

\item{stan_result}{The output of \code{GetStanSensitivityFromModelFit}}
}
\value{
A tidy sensitivity dataframe where the sensitivity is in the
direction of the difference between the hyperparameters in the two stan
data lists.
}
\description{
Use the linear approximation to predict the sensitivity to a new
stan data list.
}
