% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_helpers_lib.R
\name{TidyColumn}
\alias{TidyColumn}
\title{Convert a column of variable names to a tidy format.}
\usage{
TidyColumn(df, col, ..., variable_name = ".variable")
}
\arguments{
\item{df}{A dataframe with a parameter column to be tidied.}

\item{col}{The string name of the column to be tidied.}

\item{...}{A parameter listing in \code{tidybayes} for the parameters in the
column \code{df[[col]]}.}

\item{variable_name}{Optional.  A string for the new column's variable name.
The default is ".variable", as is standard for \code{tidybayes}.}
}
\value{
A new dataframe with extra \code{tidybayes}-style columns.
}
\description{
Convert a column of variable names to a tidy format.
}
