% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_processing_lib.R
\name{TransformSensitivityResult}
\alias{TransformSensitivityResult}
\title{Make a tidy dataframe out of a sensitivity matrix and its standard errors.}
\usage{
TransformSensitivityResult(sens_result, transform_matrix)
}
\arguments{
\item{sens_result}{The output of \code{GetStanSensitivityFromModelFit}.}

\item{transform_matrix}{A matrix where each row is a linear combination
of parameters for which to calculate sensitivity.  The row names should
describe the linear combinations.}
}
\value{
A new sensitivity result list for the specified linear combinations.
}
\description{
Make a tidy dataframe out of a sensitivity matrix and its standard errors.
}
