% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{CheckModelEquivalence}
\alias{CheckModelEquivalence}
\title{Check whether two stanfit objects agree on draws in \code{sampling_result}.}
\usage{
CheckModelEquivalence(sampling_result, model_1, model_2, stan_data_1,
  stan_data_2, check_grads = TRUE, tol = 1e-08)
}
\description{
Check whether two stanfit objects agree on draws in \code{sampling_result}.
}
