% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{EvaluateModelAtDraws}
\alias{EvaluateModelAtDraws}
\title{Evaluate a model (represented as a stanfit object) at MCMC draws, possibly
from a different model.}
\usage{
EvaluateModelAtDraws(sampling_result, model_fit, model_par_list,
  compute_grads = FALSE)
}
\arguments{
\item{sampling_result}{The output of Stan sampling (e.g. rstan::sampling())}

\item{model_fit}{A stanfit object from a model, possibly containing a
superset of the parameters in sampling_result.}

\item{compute_grads}{If FALSE, only return the log probability.}

\item{A}{list of parameters for model_fit which can be passed to
\code{unconstrain_pars} for the \code{model_fit} model.  Every parameter
in \code{get_inits} applied to \code{sampling_result} must be also be found
in \code{model_par_list}.}
}
\value{
A list with \code{lp_vec} containing a vector of log probabilities and,
if \code{compute_grads} is \code{TRUE}, gradients of the log probability,
all with resepct to the parameters in \code{model_fit} at the draws in 
\code{sampling_result}.  If the sampler contains multiple chains the
chains are concatenated in order.
}
\description{
Evaluate a model (represented as a stanfit object) at MCMC draws, possibly
from a different model.
}
