% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GenerateSensitivityFromModel}
\alias{GenerateSensitivityFromModel}
\title{Generate stan models for sensitivity calculations from a model with a
hyperparameters block.}
\usage{
GenerateSensitivityFromModel(base_model_name,
  python_script = "StanSensitivity/python/generate_models.py")
}
\arguments{
\item{base_model_name}{The name of the model with a hyperparameters block,
including the .stan suffix.}

\item{python_script}{The location of the generate_models.py from this
repository.}
}
\value{
The \code{model_name} which can be passed as an argument into other
functions in this library.  The function also generates the sensitivity model
files in the same location as the original base model.
}
\description{
Generate stan models for sensitivity calculations from a model with a
hyperparameters block.
}
\examples{
GenerateSensitivityFromModel(
    "models/my_model.stan",
    "~/git_repos/StanSensitivity/python/generate_models.py")
}
