% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sensitivity_lib.R
\name{GetSamplingModelFilename}
\alias{GetSamplingModelFilename}
\title{Get the filename of the stan model to be used for sampling.}
\usage{
GetSamplingModelFilename(model_name)
}
\arguments{
\item{model_name}{A full path to the base model name.
For example, if the model with the \code{hyperparameter} block is named
\code{/home/user/my_model.stan}, set \code{model_name} should be equal
to \code{"/home/user/my_model"}}
}
\value{
The full path of the generated file to be used for sampling.
}
\description{
Get the filename of the stan model to be used for sampling.
}
\examples{
library(devtools)
library(rstan)
rstan_options(auto_write=TRUE)

model_filename <- package_file("inst/examples/my_model.stan")
stan_data_filename <- package_file("inst/examples/my_model.data.R")

# Parse the hyperparameters block of `my_model.stan` into models that
# Stan can actuall compile.
model_name <- GenerateSensitivityFromModel(model_filename)
model <- stan_model(GetSamplingModelFilename(model_name))

# Get the original samples.
stan_data <- LoadStanData(stan_data_filename)
sampling_result <- sampling(model, data=stan_data, chains=1, iter=1000)

# Get the sensitivity result.
stan_sensitivity_list <- GetStanSensitivityModel(model_name, stan_data)
sens_result <- GetStanSensitivityFromModelFit(
    sampling_result, stan_sensitivity_list)

print(sens_result$sens_mat)
print(sens_result$sens_mat_normalized)
}
