% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_processing_lib.R
\name{GetTidyResult}
\alias{GetTidyResult}
\title{Process the results of GetStanSensitivityFromModelFit into a
tidy dataframe with standard errors.}
\usage{
GetTidyResult(sens_result)
}
\arguments{
\item{sens_result}{The output of \code{GetStanSensitivityFromModelFit}.}
}
\value{
A dataframe summarizing the sensitivity of the model posterior means
to the hyperparameters.  The reported standard errors are based on a
multivariate normal and delta method approximation which may not be
accurate for highly variable or non-normal draws. The standard errors should
be interpreted with caution.
\itemize{
    \item{hyperparameter: }{The name of the hyperparameter.}
    \item{parameter: }{The name of the parameter.}
    \item{sensitivity:}
    {The MCMC estimate of \code{dE[parameter | X, hyperparameter] /  d hyperparameter}.}
    \item{sensitivity_se: }
    {The estimated Monte Carlo error of \code{sensitivity}.}
    \item{normalized_sensitivity: }
    {The MCMC estimate of \code{sensitivity / sd(parameter}).}
    \item{normalized_sensitivity_se: }
    {The estimated Monte Carlo error of \code{normalized_sensitivity}.}
}
}
\description{
Process the results of GetStanSensitivityFromModelFit into a
tidy dataframe with standard errors.
}
